[refactor] fetch common commands and reactions to app.telegram.common
